✨ Refactor protocol generation process, include more GBIF data in it
See #189
