+++ .github/workflows/regen-arthropods-protocol.yml
@@ -24,8 +24,11 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Run crawler
-        run: npx node scripts/jessica-joachim-crawler.js
+      - name: Initialize protocol
+        run: npx node scripts/generate-protocols.js
+
+      - name: Add data from Jessica Joachim
+        run: npx node scripts/add-species-from-jessica-joachim.js
 
       - name: Set google-drive-key.json
         uses: mobiledevops/secret-to-file-action@v1
@@ -35,11 +38,14 @@ jobs:
           is-executable: false
           working-directory: ./
 
-      - name: Add species from Google Slides folder
+      - name: Add data from Google Slides folder
         run: npx node scripts/add-species-from-google-slides.js
 
       - name: Unset google-drive-key.json
         run: rm google-drive-key.json
+
+      - name: Add data from GBIF
+        run: npx node scripts/add-taxonomic-parents-from-gbif.js
 
       - name: Store scripts/jessica-joachim-crawler-report.md
         id: get-report
