--- .github/workflows/regen-taxonomy.yml
@@ -1,42 +1,0 @@
-name: Regenerate taxonomy.json
-
-on:
-  schedule:
-    # Runs at 00:00 UTC every Monday
-    - cron: '0 0 * * 1'
-  # Allow manual triggering from the Actions tab
-  workflow_dispatch:
-
-jobs:
-  generate-taxonomy:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v4
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: '22.15.0'
-          cache: 'npm'
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Generate taxonomy
-        run: npx node scripts/generate-taxonomy.js
-
-      - name: Check for changes
-        id: git-check
-        run: |
-          git diff --exit-code || echo "changes=true" >> $GITHUB_OUTPUT
-
-      - name: Commit changes if any
-        if: steps.git-check.outputs.changes == 'true'
-        run: |
-          git config --local user.email "github-actions[bot]@users.noreply.github.com"
-          git config --local user.name "github-actions[bot]"
-          git add .
-          git commit -m "🍱 Update taxonomy.json"
-          git push
