+++ scripts/add-species-from-google-slides.js
@@ -78,143 +78,147 @@ for (const { name, id } of response.data.files) {
 	const log = (...items) =>
 		console.info(`${cc.bold}${cc.cyan}[${name.padStart(longestNameLength)}]${cc.reset}`, ...items);
 
-	log('Exporting to PDF');
-	const pdf = await drive.files.export(
-		{
-			fileId: id,
-			mimeType: 'application/pdf'
-		},
-		{
-			responseType: 'arraybuffer'
-		}
-	);
-
-	log(`Converting slide 1 to image`);
-	const image = await pdfToPng(pdf.data, { pagesToProcess: [1], viewportScale: 2 })
-		.then(([{ content }]) => Jimp.fromBuffer(content))
-		.then((image) => {
-			const { width, height } = image.bitmap;
-			let cropbox = {
-				start: { x: 0, y: 0, found: false },
-				end: { x: width, y: height, found: false }
-			};
-
-			const isBackground = (x, y) => image.getPixelColor(x, y) === 0xffffffff;
-
-			image.scan((x, y) => {
-				// Get topleft corner
-				if (!cropbox.start.found && !isBackground(x, y)) {
-					cropbox.start = { x, y, found: true };
-				}
-
-				// Get bottomright corner
-				if (!cropbox.end.found && !isBackground(width - x, height - y)) {
-					cropbox.end = { x: width - x, y: height - y, found: true };
-				}
-			});
-
-			const logbox = ({ start, end }) => `(${start.x} ${start.y}) -> (${end.x} ${end.y})`;
-
-			log(
-				`Cropping image to ${cc.blue}${logbox(cropbox)}${cc.reset} from ${cc.cyan}(0 0) -> (${width} ${height})${cc.reset}`
-			);
-
-			return image.crop({
-				x: cropbox.start.x,
-				y: cropbox.start.y,
-				w: cropbox.end.x - cropbox.start.x,
-				h: cropbox.end.y - cropbox.start.y
-			});
-		});
-
-	log('Exporting as ODP');
-	const odp = await drive.files.export(
-		{
-			fileId: id,
-			mimeType: 'application/vnd.oasis.opendocument.presentation'
-		},
-		{
-			responseType: 'arraybuffer'
-		}
-	);
-
-	log('Extracting description');
-	const html = await odt
-		.toHTML({
-			path: Buffer.from(odp.data),
-			trim: true,
-			openFilters: {
-				'draw:page': ({ attributes, name }) => ({
-					xml: name,
-					html: 'section',
-					attrs: [{ name: 'data-page', value: attributes['draw:name'] }]
-				}),
-				'text:s': ({ name }) => ({
-					xml: name,
-					html: 'span',
-					attrs: [{ name: 'data-is-whitespace', value: 'true' }]
-				})
+	try {
+		log('Exporting to PDF');
+		const pdf = await drive.files.export(
+			{
+				fileId: id,
+				mimeType: 'application/pdf'
+			},
+			{
+				responseType: 'arraybuffer'
 			}
-		})
-		.then((html) => {
-			const doc = new JSDOM(html).window.document;
-			// Recover whitespace from text:s tags
-			doc.querySelectorAll('span[data-is-whitespace]').forEach((node) => {
-				node.innerHTML += ' ';
-			});
-			return doc;
-		});
-
-	const description = htmlToMarkdown(html.querySelector('section[data-page="page2"]'));
-
-	const links = Object.fromEntries(
-		[...html.querySelectorAll('section[data-page="page2"] a')]
-			.map((node) => [node.href, node.textContent])
-			.filter(([, text]) => text && text !== ' ')
-	);
-
-	const learnMore =
-		Object.entries(links).find(([, text]) => ['INPN', 'LMDI'].includes(text))[0] ?? links[0][0];
-
-	for (const [filepath, protocol] of Object.entries(protocols)) {
-		log(`Adding to ${cc.blue}${protocol.id}${cc.reset}`);
-
-		const options = protocol.metadata[`${protocol.id}__species`].options;
-
-		const imagePath = path.join(filepath.replace('.json', '.images'), `${name}.png`);
-
-		const imageUrl = new URL(
-			'https://raw.githubusercontent.com/cigaleapp/cigale/main/' +
-				path.relative(path.join(filepath, '../..'), imagePath).replaceAll('\\', '/')
 		);
 
-		if (protocol.version) {
-			imageUrl.searchParams.set('v', protocol.version);
-		}
+		log(`Converting slide 1 to image`);
+		const image = await pdfToPng(pdf.data, { pagesToProcess: [1], viewportScale: 2 })
+			.then(([{ content }]) => Jimp.fromBuffer(content))
+			.then((image) => {
+				const { width, height } = image.bitmap;
+				let cropbox = {
+					start: { x: 0, y: 0, found: false },
+					end: { x: width, y: height, found: false }
+				};
 
-		log(`Writing image to ${cc.blue}${imagePath}${cc.reset}`);
-		image.write(imagePath);
+				const isBackground = (x, y) => image.getPixelColor(x, y) === 0xffffffff;
 
-		if (options.some((o) => o.label === name)) {
-			const option = options.find((o) => o.label === name);
-			option.label = name;
-			option.learnMore = learnMore;
-			option.description = description;
-			option.links = links;
-			option.image = imageUrl;
-		} else {
-			options.push({
-				label: name,
-				key: options.length.toString(),
-				image: imageUrl,
-				description,
-				links,
-				learnMore
+				image.scan((x, y) => {
+					// Get topleft corner
+					if (!cropbox.start.found && !isBackground(x, y)) {
+						cropbox.start = { x, y, found: true };
+					}
+
+					// Get bottomright corner
+					if (!cropbox.end.found && !isBackground(width - x, height - y)) {
+						cropbox.end = { x: width - x, y: height - y, found: true };
+					}
+				});
+
+				const logbox = ({ start, end }) => `(${start.x} ${start.y}) -> (${end.x} ${end.y})`;
+
+				log(
+					`Cropping image to ${cc.blue}${logbox(cropbox)}${cc.reset} from ${cc.cyan}(0 0) -> (${width} ${height})${cc.reset}`
+				);
+
+				return image.crop({
+					x: cropbox.start.x,
+					y: cropbox.start.y,
+					w: cropbox.end.x - cropbox.start.x,
+					h: cropbox.end.y - cropbox.start.y
+				});
 			});
-		}
 
-		log(`Writing protocol to ${cc.blue}${filepath}${cc.reset}`);
-		writeFileSync(filepath, JSON.stringify(protocol, null, 2));
+		log('Exporting as ODP');
+		const odp = await drive.files.export(
+			{
+				fileId: id,
+				mimeType: 'application/vnd.oasis.opendocument.presentation'
+			},
+			{
+				responseType: 'arraybuffer'
+			}
+		);
+
+		log('Extracting description');
+		const html = await odt
+			.toHTML({
+				path: Buffer.from(odp.data),
+				trim: true,
+				openFilters: {
+					'draw:page': ({ attributes, name }) => ({
+						xml: name,
+						html: 'section',
+						attrs: [{ name: 'data-page', value: attributes['draw:name'] }]
+					}),
+					'text:s': ({ name }) => ({
+						xml: name,
+						html: 'span',
+						attrs: [{ name: 'data-is-whitespace', value: 'true' }]
+					})
+				}
+			})
+			.then((html) => {
+				const doc = new JSDOM(html).window.document;
+				// Recover whitespace from text:s tags
+				doc.querySelectorAll('span[data-is-whitespace]').forEach((node) => {
+					node.innerHTML += ' ';
+				});
+				return doc;
+			});
+
+		const description = htmlToMarkdown(html.querySelector('section[data-page="page2"]'));
+
+		const links = Object.fromEntries(
+			[...html.querySelectorAll('section[data-page="page2"] a')]
+				.map((node) => [node.href, node.textContent])
+				.filter(([, text]) => text && text !== ' ')
+		);
+
+		const learnMore =
+			Object.entries(links).find(([, text]) => ['INPN', 'LMDI'].includes(text))[0] ?? links[0][0];
+
+		for (const [filepath, protocol] of Object.entries(protocols)) {
+			log(`Adding to ${cc.blue}${protocol.id}${cc.reset}`);
+
+			const options = protocol.metadata[`${protocol.id}__species`].options;
+
+			const imagePath = path.join(filepath.replace('.json', '.images'), `${name}.png`);
+
+			const imageUrl = new URL(
+				'https://raw.githubusercontent.com/cigaleapp/cigale/main/' +
+					path.relative(path.join(filepath, '../..'), imagePath).replaceAll('\\', '/')
+			);
+
+			if (protocol.version) {
+				imageUrl.searchParams.set('v', protocol.version);
+			}
+
+			log(`Writing image to ${cc.blue}${imagePath}${cc.reset}`);
+			image.write(imagePath);
+
+			if (options.some((o) => o.label === name)) {
+				const option = options.find((o) => o.label === name);
+				option.label = name;
+				option.learnMore = learnMore;
+				option.description = description;
+				option.links = links;
+				option.image = imageUrl;
+			} else {
+				options.push({
+					label: name,
+					key: options.length.toString(),
+					image: imageUrl,
+					description,
+					links,
+					learnMore
+				});
+			}
+
+			log(`Writing protocol to ${cc.blue}${filepath}${cc.reset}`);
+			writeFileSync(filepath, JSON.stringify(protocol, null, 2));
+		}
+	} catch (error) {
+		log('An error occured:', error);
 	}
 }
